pytest/demoSimple/requirements.txt
httpx
pydantic
jsonschema
pytest
allure-pytest
python-dotenv
autopep8
